Tutorial finished !

+++ controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require('../models/Thing');
+const fs = require('fs');
 
 exports.createThing = (req, res, next) => {
     const thingObject = JSON.parse(req.body.thing);
@@ -13,25 +14,50 @@ exports.createThing = (req, res, next) => {
     thing.save()
       .then(() => res.status(201).json({ message : "Objet ajouté avec succès !"}))
       .catch(error => res.status(400).json({ error }));
-    // delete req.body._id;
-    // const thing = new Thing({
-    //   ...req.body
-    // });
-    // thing.save()
-    // .then(() => res.status(201).json({ message: "objet enregistré !"}))
-    // .catch(error => res.status(400).json({ error }));
 };
 
 exports.modifyThing = (req, res, next) => {
-    Thing.updateOne({ _id : req.params.id}, { ... req.body, _id: req.params.id})
-      .then(() => res.status(200).json({ message: "Objet modifié !"}))
+    const thingObject = req.file ? {
+      ...JSON.parse(req.body.thing),
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ... req.body };
+    delete thingObject._userId;
+    Thing.findOne({ _id : req.params.id})
+      .then(thing => {
+        if(thing.userId !== req.auth.userId){
+          res.status(401).json({ message : "Non-authorisé" });
+        } else {
+          Thing.updateOne({ _id : req.params.id }, { ...thingObject, _id : req.params.id })
+            .then(() => res.status(200).json({ message : "Objet modifié !"}))
+            .catch(error => res.status(400).json({ error }));
+        }
+      })
       .catch(error => res.status(400).json({ error }));
+    // Thing.updateOne({ _id : req.params.id}, { ... req.body, _id: req.params.id})
+    //   .then(() => res.status(200).json({ message: "Objet modifié !"}))
+    //   .catch(error => res.status(400).json({ error }));
 };
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({ _id : req.params.id})
-      .then(() => res.status(200).json({ message: 'objet supprimé ! '}))
-      .catch(error => res.status(400).json({ error }));
+    Thing.findOne({ _id : req.params.id})
+      .then(thing => {
+        if(thing.userId !== req.auth.userId){
+          res.status(401).json({ message : "Non-authorisé !"})
+        } else {
+          const filename = thing.imageUrl.split('/images/')[1]
+          fs.unlink(`images/${filename}`, () => {
+            Thing.deleteOne({ _id : req.params.id })
+              .then(() => res.status(200).json({ message : 'Objet supprimé !' }))
+              .catch(error => res.status(400).json({ error }));
+          })
+        }
+      })
+      .catch(error =>
+        res.status(500).json({ error })
+      )
+    // Thing.deleteOne({ _id : req.params.id})
+    //   .then(() => res.status(200).json({ message: 'objet supprimé ! '}))
+    //   .catch(error => res.status(400).json({ error }));
 };
 
 exports.getOneThing = (req, res, next) => {
